+++ packages/mini_program_ui/lib/src/widgets/widget_props.dart
@@ -19,6 +19,9 @@ const Set<String> mpIconNames = <String>{
   'mail',
   'home',
   'search',
+  'history',
+  'backspace',
+  'arrowBack',
 };
 
 const Set<String> mpAlignmentNames = <String>{